fix(auto-reply): share message-tool delivery hints
